requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.2
python-dotenv>=1.0
